refactor: distribute auth schemas and types into dedicated login.ts and register.ts
