+++ packages/shared/src/schemas/auth/index.ts
@@ -1,0 +1,2 @@
+export * from "./login"
+export * from "./register"
