+++ packages/shared/src/schemas/auth/login.ts
@@ -1,0 +1,45 @@
+import z from "zod"
+import { UserSchema } from "../user"
+
+export const JWTEncryptedUserSchema = z.object({
+  user: UserSchema,
+  accessToken: z.string().optional(),
+})
+
+export const UserInfoResponseSchema = z.object({
+  /**
+   * The unique ID of a Google user
+   * @example 111111111111111111111
+   */
+  sub: z.string(),
+  /**
+   * The full name of the Google user
+   * @example Straight Zhao
+   */
+  name: z.string(),
+  /**
+   * The user's first name
+   * @example Straight
+   */
+  given_name: z.string(),
+  /**
+   * The user's first name
+   * @example Zhao
+   */
+  family_name: z.string(),
+  /**
+   * The user's profile picture URL
+   */
+  picture: z.string(),
+  /**
+   * The user's email address
+   * @example straightzhao@gmail.com
+   */
+  email: z.string().email(),
+  email_verified: z.boolean(),
+  /**
+   * The hosted domain that the Google account is associated with
+   * @example aucklanduni.ac.nz
+   */
+  hd: z.string(),
+})
