+++ packages/shared/src/schemas/auth/register.ts
@@ -1,67 +1,6 @@
 import z from "zod"
-import type { User } from "../payload-types"
-import { Gender, PlayLevel, University } from "../types"
-import { MediaSchema } from "./media"
-import { CommonResponse } from "./response"
-
-// Payload User Schema
-export const UserSchema = z.object({
-  id: z.string(),
-  firstName: z.string(),
-  lastName: z.string().nullable().optional(),
-  email: z.string().email(),
-  // Payload generates a hard coded role type, the `satisfies` operator is used to ensure the type matches
-  role: z.enum(["admin", "member", "casual"]),
-  remainingSessions: z.number().nullable().optional(),
-  image: z.union([z.string(), MediaSchema]).nullable().optional(),
-  updatedAt: z.string(),
-  createdAt: z.string(),
-}) satisfies z.ZodType<User>
-
-// Google Authentication Payload JWT Schema
-export const JWTEncryptedUserSchema = z.object({
-  user: UserSchema,
-  accessToken: z.string().optional(),
-})
-
-// Google User Info Schema
-export const UserInfoResponseSchema = z.object({
-  /**
-   * The unique ID of a Google user
-   * @example 111111111111111111111
-   */
-  sub: z.string(),
-  /**
-   * The full name of the Google user
-   * @example Straight Zhao
-   */
-  name: z.string(),
-  /**
-   * The user's first name
-   * @example Straight
-   */
-  given_name: z.string(),
-  /**
-   * The user's first name
-   * @example Zhao
-   */
-  family_name: z.string(),
-  /**
-   * The user's profile picture URL
-   */
-  picture: z.string(),
-  /**
-   * The user's email address
-   * @example straightzhao@gmail.com
-   */
-  email: z.string().email(),
-  email_verified: z.boolean(),
-  /**
-   * The hosted domain that the Google account is associated with
-   * @example aucklanduni.ac.nz
-   */
-  hd: z.string(),
-})
+import { Gender, PlayLevel, University } from "../../types"
+import { CommonResponse } from "../response"
 
 // TODO: sort out discrepancy between this schema and RegisterPanelDetailsSchema
 export const RegisterDetailsSchema = z.object({
@@ -253,8 +192,3 @@ export const LoginResponseSchema = CommonResponse.extend({
    */
   data: z.string().optional(),
 })
-
-export type BasicInfoForm1Values = z.infer<typeof BasicInfoForm1Schema>
-export type BasicInfoForm2Values = z.infer<typeof BasicInfoForm2Schema>
-export type UniversityInfoFormValues = z.infer<typeof UniversityInfoFormSchema>
-export type AdditionalInfoFormValues = z.infer<typeof AdditionalInfoFormSchema>
