+++ packages/shared/src/types/auth/index.ts
@@ -1,0 +1,2 @@
+export * from "./login"
+export * from "./register"
