+++ packages/shared/src/types/auth/login.ts
@@ -4,9 +4,8 @@ import type {
   LoginFormDataSchema,
   LoginRequestBodySchema,
   LoginResponseSchema,
-  RegisterPanelDetailsSchema,
   UserInfoResponseSchema,
-} from "../schemas"
+} from "../../schemas"
 
 export const AUTH_COOKIE_NAME = "auth_token"
 export const STATE_COOKIE_NAME = "state"
@@ -15,6 +14,5 @@ export const TOKEN_EXPIRY_TIME = "1h"
 export type JWTEncryptedUser = z.infer<typeof JWTEncryptedUserSchema>
 export type UserInfoResponse = z.infer<typeof UserInfoResponseSchema>
 export type LoginRequestBody = z.infer<typeof LoginRequestBodySchema>
-export type RegisterPanelDetails = z.infer<typeof RegisterPanelDetailsSchema>
 export type LoginResponse = z.infer<typeof LoginResponseSchema>
 export type LoginFormData = z.infer<typeof LoginFormDataSchema>
